+++ src/firestore/collection/collection.spec.ts
@@ -7,6 +7,7 @@ import { QueryFn } from '../interfaces';
 
 import * as firebase from 'firebase/app';
 import { Observable } from 'rxjs/Observable';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { of } from 'rxjs/observable/of';
 import { Subscription } from 'rxjs/Subscription';
 import 'rxjs/add/operator/skip';
@@ -73,6 +74,60 @@ describe('AngularFirestoreCollection', () => {
       });
   
     });
+
+    it('should handle multiple subscriptions (hot)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.valueChanges();
+      const sub = changes.subscribe(() => {}).add(
+        changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+          sub.unsubscribe();
+        })
+      ).add(() => {
+        deleteThemAll(names, ref).then(done).catch(done.fail);
+      });
+    });
+
+    it('should handle multiple subscriptions (warm)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.valueChanges();
+      changes.take(1).subscribe(() => {}).add(() => {
+        const sub = changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+        }).add(() => {
+          deleteThemAll(names, ref).then(done).catch(done.fail);
+        });
+      });
+    });
+    
+    it('should handle dynamic queries that return empty sets', async (done) => {
+      const ITEMS = 10;
+      let count = 0;
+      let firstIndex = 0;
+      let pricefilter$ = new BehaviorSubject<number|null>(null);
+      const randomCollectionName = randomName(afs.firestore);
+      const ref = afs.firestore.collection(`${randomCollectionName}`);
+      let names = await createRandomStocks(afs.firestore, ref, ITEMS);
+      const sub = pricefilter$.switchMap(price => {
+        return afs.collection(randomCollectionName, ref => price ? ref.where('price', '==', price) : ref).valueChanges()
+      }).subscribe(data => {
+        count = count + 1;
+        // the first time should all be 'added'
+        if(count === 1) {
+          expect(data.length).toEqual(ITEMS);
+          pricefilter$.next(-1);
+        }
+        // on the second round, we should have filtered out everything
+        if(count === 2) {
+          expect(data.length).toEqual(0);
+          sub.unsubscribe();
+          deleteThemAll(names, ref).then(done).catch(done.fail);
+        }
+      });
+    });
+
   });
 
   describe('snapshotChanges()', () => {
@@ -83,7 +138,6 @@ describe('AngularFirestoreCollection', () => {
       const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
       const sub = stocks.snapshotChanges().subscribe(data => {
         const ids = data.map(d => d.payload.doc.id);
-        debugger;
         count = count + 1;
         // the first time should all be 'added'
         if(count === 1) {
@@ -99,6 +153,33 @@ describe('AngularFirestoreCollection', () => {
           sub.unsubscribe();
           deleteThemAll(names, ref).then(done).catch(done.fail);
         }
+      });
+    });
+
+    it('should handle multiple subscriptions (hot)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.snapshotChanges();
+      const sub = changes.subscribe(() => {}).add(
+        changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+          sub.unsubscribe();
+        })
+      ).add(() => {
+        deleteThemAll(names, ref).then(done).catch(done.fail);
+      });
+    });
+
+    it('should handle multiple subscriptions (warm)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.snapshotChanges();
+      changes.take(1).subscribe(() => {}).add(() => {
+        const sub = changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+        }).add(() => {
+          deleteThemAll(names, ref).then(done).catch(done.fail);
+        });
       });
     });
 
@@ -133,7 +214,7 @@ describe('AngularFirestoreCollection', () => {
       const ITEMS = 10;
       const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
       
-      const sub = stocks.snapshotChanges(['modified']).subscribe(data => {
+      const sub = stocks.snapshotChanges(['modified']).skip(1).subscribe(data => {
         sub.unsubscribe();
         const change = data.filter(x => x.payload.doc.id === names[0])[0];
         expect(data.length).toEqual(1);
@@ -161,6 +242,35 @@ describe('AngularFirestoreCollection', () => {
       });
   
       
+      names = names.concat([nextId]);
+      delayAdd(stocks, nextId, { price: 2 });
+    });
+
+    it('should be able to filter snapshotChanges() types - added/modified', async (done) => {
+      const ITEMS = 10;
+      let { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const nextId = ref.doc('a').id;
+      let count = 0;
+
+      const sub = stocks.snapshotChanges(['added', 'modified']).skip(1).take(2).subscribe(data => {
+        count += 1;
+        if (count == 1) {
+          const change = data.filter(x => x.payload.doc.id === nextId)[0];
+          expect(data.length).toEqual(ITEMS + 1);
+          expect(change.payload.doc.data().price).toEqual(2);
+          expect(change.type).toEqual('added');
+          delayUpdate(stocks, names[0], { price: 2 });
+        }
+        if (count == 2) {
+          const change = data.filter(x => x.payload.doc.id === names[0])[0];
+          expect(data.length).toEqual(ITEMS + 1);
+          expect(change.payload.doc.data().price).toEqual(2);
+          expect(change.type).toEqual('modified');
+        }
+      }).add(() => {
+        deleteThemAll(names, ref).then(done).catch(done.fail);
+      });
+
       names = names.concat([nextId]);
       delayAdd(stocks, nextId, { price: 2 });
     });
@@ -221,6 +331,33 @@ describe('AngularFirestoreCollection', () => {
           expect(data[0].type).toEqual('modified');
           deleteThemAll(names, ref).then(done).catch(done.fail);
         }
+      });
+    });
+
+    it('should handle multiple subscriptions (hot)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.stateChanges();
+      const sub = changes.subscribe(() => {}).add(
+        changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+          sub.unsubscribe();
+        })
+      ).add(() => {
+        deleteThemAll(names, ref).then(done).catch(done.fail);
+      });
+    });
+
+    it('should handle multiple subscriptions (warm)', async (done: any) => {
+      const ITEMS = 4;
+      const { randomCollectionName, ref, stocks, names } = await collectionHarness(afs, ITEMS);
+      const changes = stocks.stateChanges();
+      changes.take(1).subscribe(() => {}).add(() => {
+        const sub = changes.take(1).subscribe(data => {
+          expect(data.length).toEqual(ITEMS);
+        }).add(() => {
+          deleteThemAll(names, ref).then(done).catch(done.fail);
+        });
       });
     });